# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TLOZ Data API
  version: 2.0.0-draft
  description: |
    Document-first API for TLOZ Projects, Missions and Inventory.
    The v1 routes remain available temporarily as a compatibility bridge.

    Authentication is via Bearer token (agent API key) or browser session cookie.
    If both are present, the Bearer token takes priority.
  contact:
    name: TLOZ
  license:
    name: Proprietary
    identifier: LicenseRef-Zivelo-Proprietary
servers:
  - url: /api/v1
    description: Legacy v1 compatibility API
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: List TLOZ projects
      tags: [Projects]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: ownerId
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
      responses:
        "200":
          description: Matching projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createProject
      summary: Create a TLOZ project
      description: name is required; all other fields are optional.
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                descriptionDetail:
                  type: string
                icon:
                  type: string
                color:
                  type: string
                  pattern: "^#[0-9a-fA-F]{6}$"
                status:
                  $ref: "#/components/schemas/ProjectStatus"
                type:
                  type: string
                ownerId:
                  type: string
                startDate:
                  type: string
                  format: date
                dueDate:
                  type: string
                  format: date
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /missions:
    get:
      operationId: listMissions
      summary: List or find missions
      description: Supports pagination, text search on title, and filtering by project, owner, status, season, and episode.
      tags: [Missions]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: projectId
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/MissionStatus"
        - name: seasonId
          in: query
          schema:
            type: string
        - name: episodeId
          in: query
          schema:
            type: string
        - name: title
          in: query
          description: Case-insensitive partial text search on title.
          schema:
            type: string
      responses:
        "200":
          description: Matching missions with hydrated relations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createMission
      summary: Create a mission
      description: title and type are required. Omitted ownerId, projectId, and status default to Zibot, the Zivelo project, and later.
      tags: [Missions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type]
              properties:
                id:
                  type: string
                  description: Optional custom ID (auto-generated if omitted).
                title:
                  type: string
                description:
                  type: string
                  default: ""
                  maxLength: 280
                  description: Short mission summary shown in previews.
                descriptionDetail:
                  type: string
                  default: ""
                  maxLength: 20000
                  description: Full Markdown detail; task items feed checklist metrics.
                icon:
                  type: string
                  default: Sword
                type:
                  $ref: "#/components/schemas/MissionType"
                status:
                  $ref: "#/components/schemas/MissionStatus"
                  default: later
                ownerId:
                  type: string
                  description: Defaults to the authenticated developer for Full Stack Developer callers, otherwise to the Zibot user ID.
                projectId:
                  type: string
                  description: Defaults to the project whose slug is zivelo when omitted.
                seasonId:
                  type: string
                episodeId:
                  type: string
                dueDate:
                  type: string
                  format: date
                startDate:
                  type: string
                  format: date
                completedAt:
                  type: string
                  format: date-time
                blockedReason:
                  type: string
                progress:
                  type: integer
                  minimum: 0
                  maximum: 100
                  default: 0
                dependencyIds:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
                requiredQuestItemIds:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
                resources:
                  type: array
                  items:
                    $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Mission created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/MissionDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /quest-items:
    get:
      operationId: listQuestItems
      summary: List or find quest items
      tags: [Quest Items]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: ownerId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/InventoryStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/InventoryCategory"
      responses:
        "200":
          description: Matching quest items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestItemListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createQuestItem
      summary: Create a quest item
      description: name is required; all other fields are optional.
      tags: [Quest Items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                descriptionDetail:
                  type: string
                icon:
                  type: string
                status:
                  $ref: "#/components/schemas/InventoryStatus"
                category:
                  $ref: "#/components/schemas/InventoryCategory"
                ownerId:
                  type: string
                acquiredAt:
                  type: string
                  format: date
      responses:
        "201":
          description: Quest item created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/QuestItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /documents:
    servers:
      - url: /api/v2
    get:
      operationId: listDocuments
      summary: List and search documents
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/DocumentKind"
        - name: parent
          in: query
          description: Internal UUID or public Project ID.
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cursor-paginated documents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createDocument
      summary: Create a Project, Mission or Inventory document
      tags: [Documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentCreate"
          text/markdown:
            schema:
              type: string
      responses:
        "201":
          description: Document created.
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
            text/markdown:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /documents/{documentId}/contract:
    servers:
      - url: /api/v2
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      operationId: getProjectContract
      summary: Get a Project child-document contract
      tags: [Documents]
      responses:
        "200":
          description: Project contract.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: replaceProjectContract
      summary: Replace active fields while preserving retired values
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fields]
              properties:
                fields:
                  type: array
                  items:
                    $ref: "#/components/schemas/FieldDefinition"
      responses:
        "200":
          description: Updated Project document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The supplied revision is stale.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "428":
          description: If-Match is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: Existing NextAuth browser session. Production may use the secure cookie prefix.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: tloz_
      description: |
        New agent API keys use `tloz_` followed by 64 hex characters.
        Legacy `zaf_` keys remain valid during the v1 grace release.
        To authenticate, include the header:
        ```
        Authorization: Bearer tloz_<64-hex-chars>
        ```
        If both Bearer token and session cookie are present, the Bearer token takes priority.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        minLength: 1
...
    DocumentId:
      name: documentId
      in: path
      required: true
      description: Internal UUID or stable public ID such as TLO-0023.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Quoted revision ETag returned by the previous read.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid path or query input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: An authenticated session or valid API key is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: The authenticated actor is not permitted to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: Unexpected server error. Details are recorded server-side.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    DocumentKind:
      type: string
      enum: [project, mission, inventory]
...
    FieldDefinition:
      type: object
      additionalProperties: false
      required: [key, label, type, required, visible, position, options]
      properties:
        id:
          type: string
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, boolean, date, select, multiselect, person, relation]
        required:
          type: boolean
        visible:
          type: boolean
        position:
          type: integer
          minimum: 0
        defaultValue:
          $ref: "#/components/schemas/DocumentScalar"
        options:
          type: array
          items:
            $ref: "#/components/schemas/FieldOption"
...
    DocumentResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Document"
    DocumentListResponse:
      type: object
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        nextCursor:
          oneOf:
            - type: string
            - type: "null"
    DocumentCreate:
      type: object
      additionalProperties: false
      required: [kind, title]
      properties:
        kind:
          $ref: "#/components/schemas/DocumentKind"
        parentPublicId:
          type: string
        title:
          type: string
        summary:
          type: string
        body:
          type: string
        properties:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DocumentScalar"
        contract:
          description: Mission field contract. Only valid when kind is project.
          type: object
          additionalProperties: false
          required: [fields]
          properties:
            fields:
              type: array
              items:
                $ref: "#/components/schemas/FieldDefinition"
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum: [INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR, DOCUMENT_INVALID, DOCUMENT_NOT_FOUND, DOCUMENT_REVISION_CONFLICT, PRECONDITION_REQUIRED, INVALID_ATTACHMENT_MANIFEST, ATTACHMENT_STORAGE_NOT_CONFIGURED, ATTACHMENT_STORAGE_ERROR, attachment_object_missing, attachment_object_invalid, batch_superseded]
            message:
              type: string
            fields:
              type: object
              description: Field-specific guidance for INVALID_REQUEST responses.
              additionalProperties:
                type: string
            requestId:
              type: string

    # ── Users ────────────────────────────────────────────────────────
...
    ProjectStatus:
      type: string
      enum: [planned, active, archived]
    Project:
      type: object
      additionalProperties: false
      required: [id, slug, name, description, color, icon, status, ownerId, startDate]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        descriptionDetail:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
        icon:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        ownerId:
          type: string
        startDate:
          type: string
          format: date
        dueDate:
          type: [string, "null"]
          format: date
    ProjectListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        nextCursor:
          type: [string, "null"]

    # ── Missions ─────────────────────────────────────────────────────
    MissionStatus:
      type: string
      enum: [now, next, later, completed, blocked]
    MissionType:
      type: string
      enum: [main_quest, side_quest, farming_quest, exploration_quest]
    Mission:
      type: object
      additionalProperties: false
      required: [id, displayId, title, description, descriptionDetail, icon, type, status, ownerId, progress]
      properties:
        id:
          type: string
        displayId:
          type: string
        title:
          type: string
        description:
          type: string
          maxLength: 280
        descriptionDetail:
          type: string
          maxLength: 20000
        icon:
          type: string
        type:
          $ref: "#/components/schemas/MissionType"
        status:
          $ref: "#/components/schemas/MissionStatus"
        ownerId:
          type: string
        projectId:
          type: string
        seasonId:
          type: string
        episodeId:
          type: string
        dueDate:
          type: string
          format: date
        progress:
          type: integer
          minimum: 0
          maximum: 100
        completedAt:
          type: string
          format: date-time
        blockedReason:
          type: string
    MissionDetail:
      allOf:
        - $ref: "#/components/schemas/Mission"
        - type: object
          required: [checklist, checklistCount, completed, resources]
          properties:
            project:
              $ref: "#/components/schemas/Project"
            dependencies:
              type: array
              items:
                $ref: "#/components/schemas/Mission"
            questItems:
              type: array
              items:
                $ref: "#/components/schemas/QuestItem"
            checklist:
              type: array
              items:
                $ref: "#/components/schemas/ChecklistItem"
            checklistCount:
              type: integer
              minimum: 0
            completed:
              type: integer
              minimum: 0
            resources:
              type: array
              items:
                $ref: "#/components/schemas/Resource"
    MissionListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MissionDetail"
        nextCursor:
          type: [string, "null"]
...
    InventoryStatus:
      type: string
      enum: [locked, unlocked]
    InventoryCategory:
      type: string
      enum: [tool, access, asset, document, other]
    QuestItem:
      type: object
      additionalProperties: false
      required: [id, name, description, icon, status, category]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        status:
          $ref: "#/components/schemas/InventoryStatus"
        category:
          $ref: "#/components/schemas/InventoryCategory"
        ownerId:
          type: string
        acquiredAt:
          type: string
          format: date
    QuestItemListResponse:
      type: object
      additionalProperties: false
      required: [data, nextCursor]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/QuestItem"
        nextCursor:
          type: [string, "null"]

    # ── Resources ────────────────────────────────────────────────────
...
    ResourceType:
      type: string
      enum: [link, document, image, file, note]
    ResourceInput:
      type: object
      additionalProperties: false
      required: [type, title]
      properties:
        type:
          $ref: "#/components/schemas/ResourceType"
        title:
          type: string
        url:
          type: string
          format: uri
        fileId:
          type: string
        icon:
          type: string
...